fonctionnelle voir insert commande(date) livre et fournisseur ok

+++ 3_admin/Views/view_all_livres.php
@@ -31,7 +31,7 @@
 				<td class="td"> <?= $l->prix ?> </td>
 				<td class="td"> <?= $l->langue ?> </td>
 				<td><a href="?controller=livre&action=update_livre&id=<?= $l->id ?>"><i class="fa-solid fa-pen"></i></a></td>
-				<td class='trash'><a href='?controller=livre&action=delete_livre&id=<?= $l->id ?>' style='color: red;'><i class='fa fa-trash'></i></a></td>
+				<td class='trash'><a href='?controller=livre&action=delete_livre&id=<?= $l->id ?>' style='color: red;' onclick="return confirm('Êtes-vous sûr de vouloir supprimer ce livre ?')"><i class='fa fa-trash'></i></a></td>
 			</tr>
 		<?php endforeach; ?>
 	</tbody>
